feat: add comments container
